requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==23.1
progress==1.6
py3-pinterest==1.3.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
